Create README.md
Full stack development of a health and wellness web app for airmen/women in the Department of Defense (DoD) using Java Spring Boot to develop a REST API on the backend, React.js for the frontend, with a MySQL database. Built the frontend and backend for various features such as user registration and login, as well as community pages where users can friend other users, create groups and events, and post in these groups to foster a community around health and wellness in the Air Force. Developed an algorithm that recommends health and wellness strategies and best practices given an airman/woman’s current physical, nutritional, and mental health management skills which are assessed by various surveys. I also used my psychology background to create a personality questionnaire that assesses each airman/woman's personality based on the “Big 5” personality model to influence the health and wellness recommendations given to airmen/women by our algorithm.